python_backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==3.12.0
python-dotenv==0.19.0
bcrypt==3.2.0
python-jwt==3.3.0